Fixed bug in .fetchReqs
fixed a bug in the requirements fetcher. I think this bug would also
affect the login file reader, but it sounded like you guys had tested
that module and cleared it already so i chose not to change it. I
commented in my warning and thoughts in the event the error does arise
so we may easily fix it if needed